oscar_repo Yá se puede registrarse y popular con Axios a la BD, Subir Foto desactivada hasta tener progreso de Estiven, oscar
